Пользовательское навигационное меню SignIn-SignOut вынесено в отдельный компонент, добавлено действие Logout
