+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,6 +1,7 @@
 import {CITIES} from '../../constants';
 import {Link} from 'react-router-dom';
 import CityCards from '../../components/city-cards/city-cards';
+import UserNavigation from '../../components/user-navigation/user-navigation';
 import {offersInCity} from '../../utils';
 import {useAppSelector} from '../../hooks/index';
 
@@ -23,24 +24,7 @@ function FavoritesScreen (): JSX.Element {
                 />
               </Link>
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link to='/' className="header__nav-link header__nav-link--profile">
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                        Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
+            <UserNavigation/>
           </div>
         </div>
       </header>
