+++ project/src/pages/main-screen/main-screen.tsx
@@ -7,12 +7,11 @@ import {useAppSelector} from '../../hooks/index';
 import {offersInCity} from '../../utils';
 import SortOptions from '../../components/sort-options/sort-options';
 import LoadingOffers from '../../components/loading-offers/loading-offers';
-import {AuthorizationStatus} from '../../constants';
+import UserNavigation from '../../components/user-navigation/user-navigation';
 
 function MainScreen ():JSX.Element {
   const offers = useAppSelector((state) => state.offers);
   const city = useAppSelector((state) => state.city);
-  const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
   const isOffersDataLoading = useAppSelector((state) => state.isOffersDataLoading);
   const cityOffers = offers ? offersInCity(offers, city.name) : [];
 
@@ -32,44 +31,7 @@ function MainScreen ():JSX.Element {
                 />
               </Link>
             </div>
-            <nav className="header__nav">
-
-              {
-                authorizationStatus === AuthorizationStatus.Auth
-                  ? (
-                    <ul className="header__nav-list">
-                      <li className="header__nav-item user">
-                        <Link
-                          className="header__nav-link header__nav-link--profile"
-                          to="/favorites"
-                        >
-                          <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                          <span className="header__user-name user__name">
-                                Oliver.conner@gmail.com
-                          </span>
-                          <span className="header__favorite-count">3</span>
-                        </Link>
-                      </li>
-                      <li className="header__nav-item">
-                        <Link className="header__nav-link" to="/">
-                          <span className="header__signout">Sign out</span>
-                        </Link>
-                      </li>
-                    </ul>
-                  )
-                  : (
-                    <ul className="header__nav-list">
-                      <li className="header__nav-item user">
-                        <Link className="header__nav-link header__nav-link--profile" to="/login">
-                          <div className="header__avatar-wrapper user__avatar-wrapper">
-                          </div>
-                          <span className="header__login">Sign in</span>
-                        </Link>
-                      </li>
-                    </ul>
-                  )
-              }
-            </nav>
+            <UserNavigation/>
           </div>
         </div>
       </header>
