+++ project/src/pages/not-found-404-screen/not-found-404-screen.tsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import UserNavigation from '../../components/user-navigation/user-navigation';
 
 function NotFound404Screen ():JSX.Element {
   return (
@@ -17,27 +18,7 @@ function NotFound404Screen ():JSX.Element {
                 />
               </Link>
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link
-                    className="header__nav-link header__nav-link--profile"
-                    to="/"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                        Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
+            <UserNavigation/>
           </div>
         </div>
       </header>
