+++ project/src/pages/place-screen/place-screen.tsx
@@ -2,6 +2,7 @@ import { Link, useParams } from 'react-router-dom';
 import {MAX_PLACE_IMAGES} from '../../constants';
 import {setStarRating} from '../../utils';
 import ReviewForm from '../../components/review-form/review-form';
+import UserNavigation from '../../components/user-navigation/user-navigation';
 import {useAppSelector} from '../../hooks/index';
 
 function PlaceScreen (): JSX.Element {
@@ -26,27 +27,7 @@ function PlaceScreen (): JSX.Element {
                 />
               </Link>
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link
-                    className="header__nav-link header__nav-link--profile"
-                    to="/"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                        Oliver.conner@gmail.com
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
+            <UserNavigation/>
           </div>
         </div>
       </header>
